Add persist middleware for useHydratedCompareStore
- Divide state-util into two separate utils (compare-state and search-state)
- Add search query to "Oversikt" button in product page (breadcrump trail)
- Update zustand to newest version (v4.3.3)
