+++ package.json
@@ -33,6 +33,6 @@
     "sharp": "^0.31.2",
     "swr": "^2.0.0-rc.3",
     "typescript": "4.8.4",
-    "zustand": "^4.3.1"
+    "zustand": "^4.3.3"
   }
 }
